Worked

+++ include/pdu.h
@@ -16,7 +16,7 @@
 struct pdu {
     struct eth_hdr *ethhdr;
     struct mip_hdr *miphdr;
-    uint8_t        *sdu;
+    uint32_t        *sdu;
 } __attribute__((packed));
 
 struct pdu * alloc_pdu(void);
